+++ src/include/tensor/host_tensor_executors.hpp
@@ -2653,6 +2653,47 @@ RppStatus threshold_f16_f16_host_tensor(Rpp16f *srcPtr,
                                         RppLayoutParams layoutParams,
                                         rpp::Handle& handle);
 
+// -------------------- concat --------------------
+
+RppStatus concat_u8_u8_host_tensor(Rpp8u *srcPtr1,
+                                   Rpp8u *srcPtr2,
+                                   RpptGenericDescPtr srcPtr1GenericDescPtr,
+                                   RpptGenericDescPtr srcPtr2GenericDescPtr,
+                                   Rpp8u *dstPtr,
+                                   RpptGenericDescPtr dstGenericDescPtr,
+                                   Rpp32u axisMask,
+                                   Rpp32u *srcPtr1roiTensor,
+                                   Rpp32u *srcPtr2roiTensor,
+                                   RppLayoutParams layoutParams,
+                                   rpp::Handle& handle);
+
+RppStatus concat_f32_f32_host_tensor(Rpp32f *srcPtr1,
+                                     Rpp32f *srcPtr2,
+                                     RpptGenericDescPtr srcPtr1GenericDescPtr,
+                                     RpptGenericDescPtr srcPtr2GenericDescPtr,
+                                     Rpp32f *dstPtr,
+                                     RpptGenericDescPtr dstGenericDescPtr,
+                                     Rpp32u axisMask,
+                                     Rpp32u *srcPtr1roiTensor,
+                                     Rpp32u *srcPtr2roiTensor,
+                                     RppLayoutParams layoutParams,
+                                     rpp::Handle& handle);
+
+template<typename T1, typename T2>
+RppStatus concat_generic_host_tensor(T1 *srcPtr1,
+                                     T1 *srcPtr2,
+                                     RpptGenericDescPtr srcPtr1GenericDescPtr,
+                                     RpptGenericDescPtr srcPtr2GenericDescPtr,
+                                     T2 *dstPtr,
+                                     RpptGenericDescPtr dstGenericDescPtr,
+                                     Rpp32u axisMask,
+                                     Rpp32u *srcPtr1roiTensor,
+                                     Rpp32u *srcPtr2roiTensor,
+                                     RppLayoutParams layoutParams,
+                                     rpp::Handle& handle);
+
+// -------------------- jpeg_compression_distortion --------------------
+
 RppStatus jpeg_compression_distortion_u8_u8_host_tensor(Rpp8u *srcPtr,
                                                         RpptDescPtr srcDescPtr,
                                                         Rpp8u *dstPtr,